requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
mysqlclient
sqlalchemy
python-dotenv
python-docx
alembic
pymysql